+++ packages/uikit/src/components/hook-form/PhoneInput/index.tsx
@@ -2,7 +2,7 @@ import { SelectProps, Box, BoxProps } from '@mantine/core'
 import { createStyles } from '@mantine/core'
 import { get } from 'lodash-es'
 import { Controller, RegisterOptions, useFormContext } from 'react-hook-form'
-import { CountryData } from 'react-phone-input-2'
+import type { CountryData } from 'react-phone-input-2'
 
 import { PhoneInput, PhoneInputProps } from '../../PhoneInput'
 import { FormSelect } from '../Select'