+++ stories/biz/Form.ts
@@ -1,2 +1,0 @@
-export * from '../../packages/components/hook-form/Form'
-export * from '../../packages/components/hook-form/TextInput'